pyproject.toml
[project]
name = "GDG-Hackathon-ML6-Agent"
version = "0.1.0"
description = "This repository holds the scaffolding and dummy agent for the GDGxML6 Hackathon."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk",
    "colorama>=0.4.6",
    "pyfiglet>=1.0.4",
    "google-search-results>=2.4.2",
    "google-genai >=0.7.0",
    "PyMuPDF>=1.26.0",
    "wordninja>=2.0.0",
]
